Update README.md
Groceasy is a cross-platform grocery shopping app built using Flutter. It offers a simple and intuitive user interface for browsing grocery items, managing a shopping cart, and placing orders. Designed as a starter template, Groceasy is ideal for developers looking to build or learn how to create grocery or e-commerce applications with Flutter.